Now argv does not change. Changed read_system and write_system function declarations. clrscr() moved to system.h. Refactoring in system.c. Some changes in README.md

+++ src/vp.c
@@ -42,13 +42,13 @@ void vp_cleanup(View_Port *vp){
 	SDL_DestroyWindow(vp->screen);
 }
 
-void render_filled_circle(View_Port *vp, float x0, float y0, float r){
-	float y1 = y0, y2 = y0;
-	float x1, x2;
+void render_filled_circle(View_Port *vp, double x0, double y0, double r){
+	double y1 = y0, y2 = y0;
+	double x1, x2;
 
 	while (y1 >= y0 - r){
-		x1 = x0 - floorf(sqrtf(r*r - (y1 - y0)*(y1 - y0)));
-		x2 = x0 + floorf(sqrtf(r*r - (y1 - y0)*(y1 - y0)));
+		x1 = x0 - floor(sqrt(r*r - (y1 - y0)*(y1 - y0)));
+		x2 = x0 + floor(sqrt(r*r - (y1 - y0)*(y1 - y0)));
 
 		SDL_RenderDrawLine(vp->ren, x1, y1, x2, y1);
 		SDL_RenderDrawLine(vp->ren, x1, y2, x2, y2);
